make-dist: set version number only once
Before this commit, "git describe" was being run *twice* - once
at the beginning to initialize the version variable, and another
time further down to populate the .git_version file.

Refactor so the command is run only once, and include the --long
option to eliminate the undesirable behavior of producing just
the tag when HEAD points to a tag.

Also, since .git_version is now populated without the leading "v",
the cmake code for stripping off that leading "v" is no longer
needed.

Signed-off-by: Nathan Cutler <ncutler@suse.com>