Revert "mon/MonClient: disable ms_bind_msgr2 if NAUTILUS feature not set"
This reverts commit 396c9277628a6dd0e183202d84cfef54898ad0a2.

That behavior complicated the process of moving to v1 -> v2+v1 (extra
daemon restart, for example).  Also, it wasn't sufficient: there are lots
of other reasons besides the mon version that could lead to a mismatch
between what we bind to locally and what is published in the cluster
maps.  Better to embrace that reality and address those other issues.

Signed-off-by: Sage Weil <sage@redhat.com>