mgr/dashboard: Add config filter and delete routes
Add a filter route to the configurations endpoint to get a subset of
config options in one request.
Add a delete route to the configurations endpoint to delete a
specific config option value.

The commit contains the frontend and backend related changes.

It also adds the missing '/' to `ConfigurationService.bulkCreate` and
unit test.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>