osd: add flag to prevent truncate_seq copy in copy-from operation
It doesn't make sense to copy truncate_seq from the base object for the
CephFS use-case (when using the copy_file_range syscall).  Doing so results
in issues when performing an object copy after a truncate operation.

This patch adds a flag to the copy-from op that clients can set so that the
truncate_seq and truncate_size to set in the destination object are sent in
the op payload.

Clients can check if 'copy-from' without truncate_seq copy is supported by
checking the SERVER_OCTOPUS feature bit.

Fixes: https://tracker.ceph.com/issues/37378
Signed-off-by: Luis Henriques <lhenriques@suse.com>