vstart: add new option to pass list of block devices to bluestore
I was doing a bit of testing against a vstart cluster, and ended up
hitting ENOSPC errors. This patch adds a new --bluestore-devs option
that will set up bluestore OSDs to use a set of block devices as the
backing store instead of creating files for them.

They are assigned in order and if there are more OSDs than
bluestore-devs, then we'll create files for the remaining ones.
Also skip doing bluestore fsck on mount in this case to improve
startup times.

Signed-off-by: Jeff Layton <jlayton@redhat.com>