MERGEDOWN: repair_object/force_object_missing: simplify by using rebuild
repair_object sort of worked by accident for non-ec non-primary bad
peers.  It didn't update missing_loc, so needs_recovery() returns the
wrong answer.  However, is_unfound() does as well, so ReplicatedBackend
is nevertheless happy as the object would be present on the primary.
This patch updates the behavior to be uniform and reuses
MissingLoc.rebuild.

Signed-off-by: Samuel Just <sjust@redhat.com>