rgw_file: all directories are virtual with respect to contents
This change causes directory handles to always report an mtime of
"now."  This is not an invalidate per se--it interacts with the
nfs implementation to produce that result when the implementation
updates its cached attributes.  Hence, it can be modulated by timers
or other rules governing attribute caching at the upper layer.

Fixes: http://tracker.ceph.com/issues/40204

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>