rgw: fix checking index_doc_suffix when getting effective key
Currently, if the index_doc_suffix is empty which is leaded
by the IndexDocument field is not configurated or is set to
empty string during enabling bucket website function, the
rgw will crash when accessing the static website through the
S3Website enabled RGW instance.

In actually, we have add the necessary checking in the commit
87ccc62c0f79ea0304cef0cf7ea3046acf543d10, but double checking
in here is better.

Signed-off-by: Enming Zhang <enming.zhang@umcloud.com>