elector: Set up infrastructure for having multiple election strategies
This means we can implement a new strategy (or change an existing one) without
potentially adding bugs to the older, stable implementations. And admins
can eventually select a strategy that works for them. Right now it defaults to
CLASSIC (which I brought back after mauling it for the DISALLOW mode) and
there's no plumbing to change it in a real monitor.

Also still to-do: efficiently invoke all the unit tests on each applicable
strategy.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>