mount.ceph: clean up return codes
Mount helpers should return a specific mask as the return value so
that the caller can interpret what happened. These are detailed in
mount(8) and are already defined in mtab.c.

Fix the helper to return appropriate error codes for different
failure modes. Also, add a goto out target so we can ensure that we
do proper cleanup on failure.

Signed-off-by: Jeff Layton <jlayton@redhat.com>